% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{bump,Bumpr.Git.S3,character,character-method}
\alias{bump,Bumpr.Git.S3,character,character-method}
\title{Get Value from Environment}
\usage{
\S4method{bump}{Bumpr.Git.S3,character,character}(what, from, to,
  temp_credentials, ...)
}
\arguments{
\item{what}{\code{\link{Bumpr.Git.S3}}.}

\item{from}{\code{\link{character}}.}

\item{to}{\code{\link{character}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
\code{\link{ANY}}. Variable value
}
\description{
See generic: \code{\link[reactr]{bump}}
}
\examples{
\dontrun{

bump(what = classr::createInstance(cl = "Bumpr.Git.S3"))

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/bumpr}
}
\seealso{
\code{
   \link[reactr]{bump}
}
}

