% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{bump,Bumpr.Git.S3,missing,missing-method}
\alias{bump,Bumpr.Git.S3,missing,missing-method}
\title{Bump Version}
\usage{
\S4method{bump}{Bumpr.Git.S3,missing,missing}(what, from, to,
  temp_credentials = FALSE, ...)
}
\arguments{
\item{what}{\code{\link{Bumpr.Git.S3}}.}

\item{from}{\code{\link{missing}}.}

\item{to}{\code{\link{missing}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
See method
   \code{\link[reactr]{bump-character-character-Bumpr.Git.S3-method}}
}
\description{
See generic: \code{\link[reactr]{bump}}
}
\examples{
\dontrun{

bump(what = classr::createInstance(cl = "Bumpr.Git.S3"))

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/bumpr}
}
\seealso{
\code{
   \link[reactr]{bump}
}
}

