% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{bump,GitVersion.S3,character,character-method}
\alias{bump,GitVersion.S3,character,character-method}
\title{Bump Git Version Number}
\usage{
\S4method{bump}{GitVersion.S3,character,character}(what, from, to,
  project, temp_credentials, ...)
}
\arguments{
\item{what}{\code{\link{GitVersion.S3}}.}

\item{from}{\code{\link{character}}.}

\item{to}{\code{\link{character}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
\code{\link{list}}. (\code{old} and \code{new} package version or
   \code{list()} if the function exited before completing the bump.
}
\description{
See generic: \code{\link[reactr]{bump}}
}
\examples{
\dontrun{

##------------------------------------------------------------------------------
## DISCLAIMER  
##------------------------------------------------------------------------------

## Please that with repositories that are not crucial for your productive work!

## Make sure that you set your working directory to your package
## project's root directory and that you've read and understood the 
## assumptions that this function makes!
    
bump(what = classr::createInstance(cl = "RPackageVersion.S3"))
bump(what = classr::createInstance(cl = "GitVersion.S3"))

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/bumpr}
}
\seealso{
\code{
   \link[reactr]{bump}
}
}

