% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{bump,RPackageVersion.S3,character,character-method}
\alias{bump,RPackageVersion.S3,character,character-method}
\title{Bump Package Version Number}
\usage{
\S4method{bump}{RPackageVersion.S3,character,character}(what, from, to,
  taken = character(), desc_fields = list(Date = NULL), ...)
}
\arguments{
\item{what}{\code{\link{RPackageVersion.S3}}.}

\item{from}{\code{\link{character}}.}

\item{to}{\code{\link{character}}.}

\item{taken}{\code{\link{character}}.
Version numbers that are already taken. Usually, these are the Git
tags that correspond to release versions. If used in combination
with \code{\link{bump-Bumpr-Git.S3-character-character-method}}, this
information is automatically retrieved.}

\item{desc_fields}{\code{\link{list}}.
Additional fields (besides \code{Version}) in DESCRIPTION file that
should be updated with a version number bump. Specified as name-value pairs.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
\code{\link{list}}. (\code{old} and \code{new} package version or
   \code{list()} if the function exited before completing the bump.
}
\description{
See generic: \code{\link[reactr]{bump}}
}
\examples{
\dontrun{

##------------------------------------------------------------------------------
## DISCLAIMER  
##------------------------------------------------------------------------------

## Please that with repositories that are not crucial for your productive work!

## Make sure that you set your working directory to your package
## project's root directory and that you've read and understood the 
## assumptions that this function makes!
    
bump(what = classr::createInstance(cl = "RPackageVersion.S3"))
bump(what = classr::createInstance(cl = "GitVersion.S3"))

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/bumpr}
}
\seealso{
\code{
   \link[reactr]{bump}
}
}

