% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{bump,RPackageVersion.S3,missing,missing-method}
\alias{bump,RPackageVersion.S3,missing,missing-method}
\title{Bump Package Version Number}
\usage{
\S4method{bump}{RPackageVersion.S3,missing,missing}(what, from, to,
  temp_credentials = FALSE, ...)
}
\arguments{
\item{what}{\code{\link{RPackageVersion.S3}}.}

\item{from}{\code{\link{missing}}.}

\item{to}{\code{\link{missing}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
See method
   \code{\link[reactr]{bump-character-character-RPackageVersion.S3-method}}
}
\description{
See generic: \code{\link[reactr]{bump}}
}
\examples{
\dontrun{

##------------------------------------------------------------------------------
## DISCLAIMER  
##------------------------------------------------------------------------------

## Please that with repositories that are not crucial for your productive work!

## Make sure that you set your working directory to your package
## project's root directory and that you've read and understood the 
## assumptions that this function makes!
    
bump(what = classr::createInstance(cl = "RPackageVersion.S3"))
bump(what = classr::createInstance(cl = "GitVersion.S3"))

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/bumpr}
}
\seealso{
\code{
   \link[reactr]{bump}
}
}

