% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{bump,GitVersion.S3,character,character-method}
\alias{bump,GitVersion.S3,character,character-method}
\title{Bump Git Version Number}
\usage{
\S4method{bump}{GitVersion.S3,character,character}(what, from, to, project,
  temp_credentials, ...)
}
\arguments{
\item{what}{\code{\link{GitVersion.S3}}.}

\item{from}{\code{\link{character}}.}

\item{to}{\code{\link{character}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
\code{\link{list}}. (\code{old} and \code{new} version number and
   the git tag (\code{git_tag}) or
   \code{list()} if the function exited before completing the bump.
}
\description{
See generic: \code{\link[bumpr]{bump}}
}
\examples{
\dontrun{

##------------------------------------------------------------------------------
## DISCLAIMER  
##------------------------------------------------------------------------------

## Please that with repositories that are not crucial for your productive work!

## Make sure that you set your working directory to your package
## project's root directory and that you've read and understood the 
## assumptions that this function makes!
    
bump(what = bumpr::RPackageVersion.S3())
bump(what = bumpr::GitVersion.S3())

}
}
\author{
Janko Thyson \email{janko.thyson@gmail.com}
}
\references{
\url{http://github.com/jankothyson/bumpr}
}
\seealso{
\code{
   \link[bumpr]{bump}
}
}

