% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{bump,RPackageVersion.S3,missing,missing-method}
\alias{bump,RPackageVersion.S3,missing,missing-method}
\title{Bump Package Version Number}
\usage{
\S4method{bump}{RPackageVersion.S3,missing,missing}(what, from, to,
  desc_fields = list(Date = NULL), taken = character(), ...)
}
\arguments{
\item{what}{\code{\link{RPackageVersion.S3}}.}

\item{from}{\code{\link{missing}}.}

\item{to}{\code{\link{missing}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
See method
   \code{\link[bumpr]{bump-character-character-RPackageVersion.S3-method}}
}
\description{
See generic: \code{\link[bumpr]{bump}}.
See main method: \code{\link[bumpr]{bump-RPackageVersion.S3-character-character-method}}.
}
\examples{
\dontrun{

##------------------------------------------------------------------------------
## DISCLAIMER  
##------------------------------------------------------------------------------

## Please that with repositories that are not crucial for your productive work!

## Make sure that you set your working directory to your package
## project's root directory and that you've read and understood the 
## assumptions that this function makes!
    
bump(what = bumpr::RPackageVersion.S3())
bump(what = bumpr::GitVersion.S3())

}
}
\author{
Janko Thyson \email{janko.thyson@gmail.com}
}
\references{
\url{http://github.com/jankothyson/bumpr}
}
\seealso{
\code{
   \link[bumpr]{bump},
   \link[bumpr]{bump-RPackageVersion.S3-character-character-method}
}
}

