% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bump}
\alias{bump}
\title{Bump Something}
\usage{
bump(what, from, to, ...)
}
\arguments{
\item{what}{\strong{Signature argument}.
Object containing information about what should be bumped.}

\item{from}{\strong{Signature argument}.
Object containing current information.}

\item{to}{\strong{Signature argument}.
Object containing new information.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Bumps a ressource - whatever that means is left up to the methods of this
generic function.
}
\examples{
\dontrun{

bump(what = classr::createInstance(cl = "Bumpr.Git.S3"))

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/bumpr}
}
\seealso{
\code{
  	\link[reactr]{bump-Bumpr.Git.S3-character-character-method}
}
}

