% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{bumpGitVersion,Bumpr.GitVersion.S3-method}
\alias{bumpGitVersion,Bumpr.GitVersion.S3-method}
\title{Bump Git Version Number}
\usage{
\S4method{bumpGitVersion}{Bumpr.GitVersion.S3}(project = character(),
  temp_credentials = FALSE, .ns = NULL, ...)
}
\arguments{
\item{.ns}{\code{\link{Bumpr.GitVersion.S3}}.}

\item{project}{\code{\link{character}}.
Name of the project under Git version control.}

\item{temp_credentials}{\code{\link{logical}}.
\code{TRUE}: delete HTTPS credentials after each bump;
\code{FALSE}: permanently store HTTPS credentials in \code{_netrc} file.
See details.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
\code{\link{character}}. New Git version that the project has
   been bumped to.
}
\description{
See generic: \code{\link[reactr]{bumpGitVersion}}
}
\examples{
\dontrun{

##------------------------------------------------------------------------------
## DISCLAIMER  
##------------------------------------------------------------------------------

## Please that with repositories that are not crucial for your productive work!

## Make sure that you set your working directory to your package
## project's root directory and that you've read and understood the 
## assumptions that this function makes!
  
bumpGitVersion()

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/bumpr}
}
\seealso{
\code{
   \link[reactr]{bumpGitVersion}
}
}

