% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bumpGitVersion}
\alias{bumpGitVersion}
\title{Bump Git Version}
\usage{
bumpGitVersion(temp_credentials = FALSE, .ns = NULL, ...)
}
\arguments{
\item{temp_credentials}{\code{\link{logical}}.
\code{TRUE}: delete HTTPS credentials after each bump;
\code{FALSE}: permanently store HTTPS credentials in \code{_netrc} file.
See details.}

\item{.ns}{\strong{Signature argument}.
Object containing namespace information.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Bumps the version of a git project to the next version.
}
\details{
Asumptions:
\itemize{
  \item{\strong{R package project}: } {

     You are using this function to systematically manage the versions
     of an R package project.
  }
  \item{\strong{Local Git repository}: }{

     Your package project is under Git version control, i.e. a local Git
     repository has been created in your package project's root directory.
     (look for directory \code{.git} in your package project's root
     directory)
  }
  \item{\strong{Remote Git repository}: }{

     At the very least one remote repository with name \code{origin}
     has been defined for your local Git repository.

     Additional remote repositories with different
     names are not a problem. You can choose them interactively.
     (Run \code{git remote} in your git shell to find out about your
     remote repositories)
  }
  \item{\strong{HTTP credentials}: }{

     If you want to push to a GitHub repository or any other remote
     repository that relies on HTTPS for authentication, the function
     assumes that you \strong{are willing to store (at least temporarily) your
     HTTP credentials in this file}:
     \code{file.path(Sys.getenv("HOME")/_netrc)}.

     Currently only tested for
     GitHub repositories as this is the location where the API seems to
     expect HTTP credentials when pushing to such a repository.

     \strong{However, You can choose to destroy this file after each bump by setting
     \code{temp_credentials = TRUE}}.
     I will try to find better ways of handling HTTPS credentials in future
     releases.
  }
}
}
\section{Disclaimer}{

\strong{This package is really new. So please test this function with
repositories that are not crucial for your productive work!}
}

\section{Recommendations}{

\itemize{
  \item{\strong{Initial commit}: }{

     Make sure that you already issued an initial commit for your \strong{local}
     repository and pushed this to your remote repository.
     The function does have built-in checks for very early stages of a
     Git repository (i.e. no commits yet, not \code{.gitignore} file yet,
     no branches on the remote repository yet), but I would not consider
     this completely stable and tested yet. If you want to check out what
     the function does in such early stages, I would recommend testing it
     with a toy Git/GitHub repository first
  }
}
}
\examples{
\dontrun{

##------------------------------------------------------------------------------
## DISCLAIMER  
##------------------------------------------------------------------------------

## Please that with repositories that are not crucial for your productive work!

## Make sure that you set your working directory to your package
## project's root directory and that you've read and understood the 
## assumptions that this function makes!
  
bumpGitVersion()

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/bumpr}
}
\seealso{
\code{
  	\link[reactr]{bumpGitVersion-Bumpr.Git.S3-method}
}
}

