% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{bumpPackageVersion,RPackageVersion.S3-method}
\alias{bumpPackageVersion,RPackageVersion.S3-method}
\title{Bump Package Version Number}
\usage{
\S4method{bumpPackageVersion}{RPackageVersion.S3}(taken = character(),
  desc_fields = list(Date = NULL), .ns = NULL, ...)
}
\arguments{
\item{taken}{\code{\link{character}}.
Version numbers that are already taken. Usually, these are the Git
tags that correspond to release versions. If used in combination
with \code{\link{bump-Bumpr-Git.S3-character-character-method}}, this
information is automatically retrieved.}

\item{desc_fields}{\code{\link{list}}.
Additional fields (besides \code{Version}) in DESCRIPTION file that
should be updated with a version number bump. Specified as name-value pairs.}

\item{.ns}{\code{\link{RPackageVersion.S3}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
\code{\link{list}}. (\code{old} and \code{new} version number or
   \code{list()} if the function exited before completing the bump.
}
\description{
See generic: \code{\link[bumpr]{RPackageVersion.S3}}
}
\examples{
\dontrun{

##------------------------------------------------------------------------------
## DISCLAIMER  
##------------------------------------------------------------------------------

## Please that with repositories that are not crucial for your productive work!

## Make sure that you set your working directory to your package
## project's root directory and that you've read and understood the 
## assumptions that this function makes!
  
bumpPackageVersion()

}
}
\author{
Janko Thyson \email{janko.thyson@gmail.com}
}
\references{
\url{http://github.com/jankothyson/bumpr}
}
\seealso{
\code{
   \link[bumpr]{bumpPackageVersion},
    \link[bumpr]{bumpGitVersion}
}
}

