% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bumpPackageVersion}
\alias{bumpPackageVersion}
\title{Bump Package Version Number}
\usage{
bumpPackageVersion(taken = character(), desc_fields = list(Date = NULL),
  .ns = NULL, ...)
}
\arguments{
\item{.ns}{\strong{Signature argument}.
Object containing namespace information.}

\item{taken}{\code{\link{character}}.
Version numbers that are already taken. Usually, these are the Git
tags that correspond to release versions. If used in combination
with \code{\link{bump-Bumpr-Git.S3-character-character-method}}, this
information is automatically retrieved.}

\item{desc_fields}{\code{\link{list}}.
Additional fields (besides \code{Version}) in DESCRIPTION file that
should be updated with a version number bump. Specified as name-value pairs.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Bumps an R package project to the next version number.
}
\details{
Asumptions:
\itemize{
  \item{\strong{R package project}: } {

     You are using this function to systematically manage the versions
     of an R package project that follows the official conventions
     (see \url{http://cran.r-project.org/doc/manuals/r-release/R-exts.html}.
  }
}
}
\section{Disclaimer}{

\strong{This package is really new. So please test this function with
repositories that are not crucial for your productive work!}
}
\examples{
\dontrun{

##------------------------------------------------------------------------------
## DISCLAIMER  
##------------------------------------------------------------------------------

## Please that with repositories that are not crucial for your productive work!

## Make sure that you set your working directory to your package
## project's root directory and that you've read and understood the 
## assumptions that this function makes!
  
bumpPackageVersion()

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/bumpr}
}
\seealso{
\code{
  	\link[reactr]{bumpPackageVersion-GitVersion.S3-method}
}
}

