% Generated by roxygen2 (4.0.2): do not edit by hand
\name{GitVersion.S3}
\alias{GitVersion.S3}
\title{Class: GitVersion.S3}
\usage{
GitVersion.S3(.x, version = character(), remote_name = "origin",
  remote_url = character(), user_email = character(),
  user_name = character(), home = Sys.getenv("HOME"))
}
\arguments{
\item{.x}{\code{\link{ANY}}. An object of an arbitrary class whose class
   attribute should be updated so that it becomes an instance of class
   \code{GitVersion.S3}. Mainly intended for rapid prototyping
   purposes}
}
\value{
Instance of class \code{GitVersion.S3}.
}
\description{
Class representing git versions (S3) and its constructor function.
}
\section{Fields}{

\describe{
\item{\code{version}}{\code{\link{character}}. Version.}

\item{\code{remote_name}}{\code{\link{character}}. Name of remote repository.}

\item{\code{remote_url}}{\code{\link{character}}. URL of remote repository.}

\item{\code{user_email}}{\code{\link{character}}. User e-mail.}

\item{\code{user_name}}{\code{\link{character}}. User name.}

\item{\code{home}}{\code{\link{character}}. User home directory.}
}}
\section{Intended use of this class}{


This S3 class, or to be more precise its constructor function, exists mainly
for rapid prototyping purposes.
This is mainly reflected in the fact, that when specifying \code{.x}, this
constructor function will simply update the \code{class} attribute of
whatever object has been provided.

However, it also allows for a more formal OOP-style of rapid
prototyping by offering explicit \emph{class fields} (all arguments except
\code{.x}). Nevertheless, it is probably advisable to switch to an
explicit formal approach such as \emph{S4} and/or \emph{Reference Classes}
once the package or application has reached a certain state of maturity.
}
\examples{
\dontrun{

## Informal use (intended mainly for rapid prototyping) //
## Takes *any* object and simply changes the class attributes
GitVersion.S3("0.1.1")  
GitVersion.S3(rep("0.1.1", 3))  
GitVersion.S3(as.list(rep("0.1.1", 3)))
GitVersion.S3(TRUE)  

## Formal use (explicitly using 'fields') //
GitVersion.S3()
GitVersion.S3(
  version = "0.1.1",
  remote_name = "origin",
  remote_url = "https://github.com/Rappster/bumpr",
  user_email = "janko.thyson@rappster.de",
  user_name = "Janko Thyson",
  home = Sys.getenv("HOME")
)

## Recommended: inlcude namespace //
## Regardless if you plan on using this class in an informal or formal way
bumpr::GitVersion.S3("0.1.1")

}
}
\author{
Janko Thyson \email{janko.thyson@gmail.com}
}
\references{
\url{http://github.com/jankothyson/bumpr}
}
\seealso{
\code{
  	\link[bumpr]{PackageVersion.S3}
}
}

