% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{resetPackageVersion,character-method}
\alias{resetPackageVersion,character-method}
\title{Reset Package Version Number}
\usage{
\S4method{resetPackageVersion}{character}(vsn, desc_fields = list(), ...)
}
\arguments{
\item{vsn}{\code{\link{character}}.}

\item{desc_fields}{\code{\link{list}}.
Additional fields (besides \code{Version}) in DESCRIPTION file that
should also be reset. Specified as name-value pairs.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
\code{\link{packageDescription}}.
}
\description{
See generic: \code{\link[bumpr]{resetPackageVersion}}
}
\examples{
\dontrun{

## Careful! 
## This simulates the situation when a version bump is rolled back due to 
## 'bumpGitVersion()' being aborted or in case it failed.
## This changes the version in your DESCRIPTION file and you should make sure
## that you change it back to the correct one after trying out the function
  
vsn_before_reset <- as.list(read.dcf("DESCRIPTION")[1,])$Version

resetPackageVersion(vsn = "0.1")
vsn_after_reset <- as.list(read.dcf("DESCRIPTION")[1,])$Version

## Cleanup //
resetPackageVersion(vsn = vsn_before_reset)

}
}
\author{
Janko Thyson \email{janko.thyson@gmail.com}
}
\references{
\url{http://github.com/jankothyson/bumpr}
}
\seealso{
\code{
   \link[bumpr]{resetPackageVersion}
}
}

