% Generated by roxygen2 (4.0.2): do not edit by hand
\name{RPackageVersion.S3}
\alias{RPackageVersion.S3}
\title{Class: RPackageVersion.S3}
\usage{
RPackageVersion.S3(.x, version = character(), lib = character(),
  path = character())
}
\arguments{
\item{.x}{\code{\link{ANY}}. An object of an arbitrary class whose class
   attribute should be updated so that it becomes an instance of class
   \code{RPackageVersion.S3}. Mainly intended for rapid prototyping
   purposes}
}
\value{
Instance of class \code{RPackageVersion.S3}.
}
\description{
Class representing R package versions (S3) and its constructor function.
}
\section{Fields}{

\describe{
\item{\code{version}}{\code{\link{character}}. Version.}

\item{\code{lib}}{\code{\link{character}}. Library location.}

\item{\code{path}}{\code{\link{character}}. Path to package project.}
}}
\section{Intended use of this class}{


This S3 class, or to be more precise its constructor function, exists mainly
for rapid prototyping purposes.
This is mainly reflected in the fact, that when specifying \code{.x}, this
constructor function will simply update the \code{class} attribute of
whatever object has been provided.

However, it also allows for a more formal OOP-style of rapid
prototyping by offering explicit \emph{class fields} (all arguments except
\code{.x}). Nevertheless, it is probably advisable to switch to an
explicit formal approach such as \emph{S4} and/or \emph{Reference Classes}
once the package or application has reached a certain state of maturity.
}
\examples{
\dontrun{

## Informal use (intended mainly for rapid prototyping) //
## Takes *any* object and simply changes the class attributes
RPackageVersion.S3("0.1.1")  
RPackageVersion.S3(rep("0.1.1", 3))  
RPackageVersion.S3(as.list(rep("0.1.1", 3)))
RPackageVersion.S3(TRUE)  

## Formal use (explicitly using 'fields') //
RPackageVersion.S3()
RPackageVersion.S3(
  version = "0.1.1",
  lib = .libPaths()[1],
  path = getwd()
)

## Recommended: inlcude namespace //
## Regardless if you plan on using this class in an informal or formal way
bumpr::RPackageVersion.S3("0.1.1")

}
}
\author{
Janko Thyson \email{janko.thyson@gmail.com}
}
\references{
\url{http://github.com/jankothyson/bumpr}
}
\seealso{
\code{
  	\link[bumpr]{GitVersion.S3}
}
}

