% Generated by roxygen2 (4.0.2): do not edit by hand
\name{SystemState.S3}
\alias{SystemState.S3}
\title{Class: SystemState.S3}
\usage{
SystemState.S3(.x, ask_authentication = TRUE, branch = "master",
  cmd_user_email = "git config --global user.email",
  cmd_user_name = "git config --global user.name", description_old = list(),
  git_tag = character(), git_user_email = character(),
  git_user_name = character(), global_or_local = "global",
  pat_or_basic = character(), path_netrc = file.path(Sys.getenv("HOME"),
  "_netrc"), path_netrc_tmp = file.path(Sys.getenv("HOME"), "_netrc_0"),
  quit = FALSE, remote = character(), remote_all = list(origin =
  character()), remote_name = "origin", remote_url = character(),
  temp_credentials = FALSE, what = NA)
}
\arguments{
\item{.x}{\code{\link{ANY}}. An object of an arbitrary class whose class
   attribute should be updated so that it becomes an instance of class
   \code{SystemState.S3}. Mainly intended for rapid prototyping
   purposes}
}
\value{
Instance of class \code{SystemState.S3}.
}
\description{
Class representing the system state (S3) and its constructor function.
}
\section{Fields}{

\describe{
\item{\code{ask_authentication}}{\code{\link{logical}}.
Controls if user is asked how to handle authentication.
Default: \code{TRUE}.}

\item{\code{branch}}{\code{\link{character}}.
Name of branch to push to remote.
Default: \code{"master"}.}

\item{\code{cmd_user_email}}{\code{\link{character}}.
Git command used to query \code{user.email}.
Default: \code{"git config --global user.email"}.}

\item{\code{cmd_user_name}}{\code{\link{character}}.
Git command used to query \code{user.name}.
Default: \code{"git config --global user.name"}.}

\item{\code{description_old}}{\code{\link{character}}.
Old state of \code{DESCRIPTION} file that can be used to rollback
changes.
Default: \code{list()}.}

\item{\code{git_tag}}{\code{\link{character}}.
Git tag for new version.
Default: \code{character()}.}

\item{\code{git_user_email}}{\code{\link{character}}.
Git value for \code{user.email}.
Default: \code{character()}.}

\item{\code{git_user_name}}{\code{\link{character}}.
Git value for \code{user.name}.
Default: \code{character}.}

\item{\code{global_or_local}}{\code{\link{character}}.
Controls if global or local Git user credentials are used.
Default: \code{"global"}.}

\item{\code{pat_or_basic}}{\code{\link{logical}}.
Controls if a personal access token (PAT or OAuth token) or basic
HTTPS authentication should be used.
Default: \code{character()} which means that user is asked.}

\item{\code{path_netrc}}{\code{\link{character}}.
Path to \code{_netrc} file used for HTTPS authentication.
Default: \code{file.path(Sys.getenv("HOME"), "_netrc")}.}

\item{\code{path_netrc_tmp}}{\code{\link{character}}.
Path to temporary \code{_netrc} file used for HTTPS authentication.
Default: \code{file.path(Sys.getenv("HOME"), "_netrc_0")}.}

\item{\code{quit}}{\code{\link{logical}}.
Controls if function quits.
Default: \code{FALSE}.}

\item{\code{remote}}{\code{\link{character}}.
Name \strong{or} URL of remote repository to push to (depending on
authentication type).
Default: \code{character()}.}

\item{\code{remote_all}}{\code{\link{character}}.
Name and URL of remote repository to push to.
Default: \code{list(origin = character())}.}

\item{\code{remote_name}}{\code{\link{character}}.
Name of remote repository to push to.
Default: \code{"origin"}.}

\item{\code{remote_url}}{\code{\link{character}}.
URL of remote repository to push to.
Default: \code{character()}.}

\item{\code{temp_credentials}}{\code{\link{logical}}.
Controls if HTTPS credentials should only be stored temporarily.
Only relevant if authentication type is \code{basic} (HTTPS).
Default: \code{FALSE}.}

\item{\code{what}}{\code{\link{ANY}}.
Contains the object to bump.
Default: \code{NA}.}

\item{\code{quit}}{\code{\link{logical}}.
Controls if function quits.
Default: \code{FALSE}.}
}}
\section{Intended use of this class}{


This S3 class, or to be more precise its constructor function, exists mainly
for rapid prototyping purposes.
This is mainly reflected in the fact, that when specifying \code{.x}, this
constructor function will simply update the \code{class} attribute of
whatever object has been provided.

However, it also allows for a more formal OOP-style of rapid
prototyping by offering explicit \emph{class fields} (all arguments except
\code{.x}). Nevertheless, it is probably advisable to switch to an
explicit formal approach such as \emph{S4} and/or \emph{Reference Classes}
once the package or application has reached a certain state of maturity.
}
\examples{
\dontrun{

## Informal use (intended mainly for rapid prototyping) //
## Takes *any* object and simply changes the class attributes
SystemState.S3(TRUE)  

## Formal use (explicitly using 'fields') //
SystemState.S3()
SystemState.S3(
  ask_authentication = TRUE,
  branch = "master",
  cmd_user_email = "git config --global user.email",
  cmd_user_name = "git config --global user.name",
  description_old = as.list(read.dcf("DESCRIPTION")[1,]),
  git_tag = "v1.1.2",
  git_user_email = "janko.thyson@rappster.de",
  git_user_name = "Janko Thyson",
  global_or_local = "global",
  pat_or_basic = "pat",
  path_netrc = file.path(Sys.getenv("HOME"), "_netrc"),
  path_netrc_tmp = file.path(Sys.getenv("HOME"), "_netrc_0"),
  quit = FALSE,
  remote = "origin",
  remote_all = list(origin = character()),
  remote_name = "origin",
  remote_url = "https://github.com/Rappster/bumpr",
  temp_credentials = FALSE,
  what = bumpr::GitVersion.S3()
)

## Recommended: inlcude namespace //
## Regardless if you plan on using this class in an informal or formal way
bumpr::SystemState.S3()

}
}
\author{
Janko Thyson \email{janko.thyson@gmail.com}
}
\references{
\url{http://github.com/jankothyson/bumpr}
}
\seealso{
\code{
  	\link[bumpr]{GitVersion.S3}
    \link[bumpr]{RPackageVersion.S3}
}
}

