% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{bump,Bumpr.GitVersion.S3,missing,missing-method}
\alias{bump,Bumpr.GitVersion.S3,missing,missing-method}
\title{Bump Git Version Number}
\usage{
\S4method{bump}{Bumpr.GitVersion.S3,missing,missing}(what, from, to,
  project = character(), temp_credentials = FALSE, ...)
}
\arguments{
\item{what}{\code{\link{Bumpr.GitVersion.S3}}.}

\item{from}{\code{\link{missing}}.}

\item{to}{\code{\link{missing}}.}

\item{temp_credentials}{\code{\link{logical}}.
\code{TRUE}: delete HTTPS credentials after each bump;
\code{FALSE}: permanently store HTTPS credentials in \code{_netrc} file.
See details.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
See method
   \code{\link[reactr]{bump-character-character-Bumpr.GitVersion.S3-method}}
}
\description{
See generic: \code{\link[reactr]{bump}}
}
\examples{
\dontrun{

##------------------------------------------------------------------------------
## DISCLAIMER  
##------------------------------------------------------------------------------

## Please that with repositories that are not crucial for your productive work!

## Make sure that you set your working directory to your package
## project's root directory and that you've read and understood the 
## assumptions that this function makes!
    
bump(what = classr::createInstance(cl = "Bumpr.RPackageVersion.S3"))
bump(what = classr::createInstance(cl = "Bumpr.GitVersion.S3"))

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/bumpr}
}
\seealso{
\code{
   \link[reactr]{bump}
}
}

