% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bump}
\alias{bump}
\title{Bump Something}
\usage{
bump(what, from, to, ...)
}
\arguments{
\item{what}{\strong{Signature argument}.
Object containing information about what should be bumped.}

\item{from}{\strong{Signature argument}.
Object containing current information.}

\item{to}{\strong{Signature argument}.
Object containing new information.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Bumps a ressource - whatever that means is left up to the methods of this
generic function.
}
\examples{
\dontrun{

##------------------------------------------------------------------------------
## DISCLAIMER  
##------------------------------------------------------------------------------

## Please that with repositories that are not crucial for your productive work!

## Make sure that you set your working directory to your package
## project's root directory and that you've read and understood the 
## assumptions that this function makes!
    
bump(what = bumpr::RPackageVersion.S3())
bump(what = bumpr::GitVersion.S3())

}
}
\author{
Janko Thyson \email{janko.thyson@gmail.com}
}
\references{
\url{http://github.com/jankothyson/bumpr}
}
\seealso{
\code{
  	\link[bumpr]{bump-GitVersion.S3-character-character-method}
}
}

