% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{bumpGitVersion,missing-method}
\alias{bumpGitVersion,missing-method}
\title{Bump Git Version}
\usage{
\S4method{bumpGitVersion}{missing}(temp_credentials = FALSE, .ns = NULL,
  ...)
}
\arguments{
\item{.ns}{\code{\link{Bumpr.Git.S3}}.}

\item{temp_credentials}{\code{\link{logical}}.
\code{TRUE}: delete HTTPS credentials after each bump;
\code{FALSE}: permanently store HTTPS credentials in \code{_netrc} file.
See details.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
See method
   \code{\link[reactr]{bumpGitVersion-Bumpr.Git.S3-method}}
}
\description{
See generic: \code{\link[reactr]{bumpGitVersion}}
}
\examples{
\dontrun{

##------------------------------------------------------------------------------
## DISCLAIMER  
##------------------------------------------------------------------------------

## Please that with repositories that are not crucial for your productive work!

## Make sure that you set your working directory to your package
## project's root directory and that you've read and understood the 
## assumptions that this function makes!
  
bumpGitVersion()

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/bumpr}
}
\seealso{
\code{
   \link[reactr]{bumpGitVersion}
}
}

