% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{bumpr}
\alias{bumpr}
\alias{bumpr-package}
\title{Easy systematic version bumping and more}
\description{
Makes it really easy to bump the version of an R package project placed
under Git version control (see \url{http://git-scm.com}) to a new version
number.
}
\details{
Version numbers must follow the \emph{semantical versioning} conventions
(see \url{http://semver.org/}).
}
\section{Classes and constructors}{


\itemize{
   \item{\code{\link[bumpr]{GitVersion.S3}}: }{

     Constructor function for instances of class
     \code{\link[bumpr]{GitVersion.S3}}
   }
   \item{\code{\link[bumpr]{RPackageVersion.S3}}: }{

     Constructor function for instances of class
     \code{\link[bumpr]{RPackageVersion.S3}}
   }
   \item{\code{\link[bumpr]{SystemState.S3}}: }{

     Constructor function for instances of class
     \code{\link[bumpr]{SystemState.S3}}.
     Used to track system states across function calls.
   }
}
}

\section{Main functions/methods}{


 \itemize{
   \item{\code{\link[bumpr]{bumpPackageVersion}}: }{

     Retrieves the current package version from the \code{DESCRIPTION} file,
     suggest the next version number and prompts the user for a new
     version number. After asking permission, the new version number is
     written to the \code{DESCRIPTION} file along with additional
     information provided via \code{desc_fields}. Currently, only
     an element of form \code{Date = NULL} is allowed/used, which
     corresponds to also updating the \code{Date} field of the
     \code{DESCRIPTION} file. \code{desc_fields = list()} suppresses that.

     Essentially, this function is a mere convenience wrapper for the
     actual workhorse function \code{\link[bumpr]{bump}} and its
     method associated to class \code{Bumpr.RPackageVersion.s3}.
   }
   \item{\code{\link[bumpr]{bumpGitVersion}}: }{

     Performs all sorts of Git-related checks and tasks in order to take care
     that everything necessary is done that is related to bumping a project
     to a higher version number.

     This provided version number is transferred to \code{v{version-number}},
     e.g. \code{v0.1.1}, and added as a Git tag.
     All commits linked to the \emph{previous} version/tag are queried and added
     to file \code{CHANGES.md}. Additionally, a template section to state the
     changes in the \emph{new} version is added in file \code{NEWS.md}.

     Files \code{DESCRIPTION} and \code{CHANGES.md} are automatically
     commited to signal the version bump.

     Optionally, you can push the new version (i.e. the new tag) as well
     as the associated commit to a remote repository (default: \code{origin}).
     This can be any valid Git remote repository, including a GitHub
     repository

     Essentially, this function is a mere convenience wrapper for the
     actual workhorse function \code{\link[bumpr]{bump}} and its
     method associated to class \code{Bumpr.GitVersion.s3}.
   }
   \item{\code{\link[bumpr]{bump}}: }{

     The actual workhorse function of this package. Introduced in order to
     generalize the package's purpose and scope to other aspects than
     version bumping.

     The long-term goal of this package is to provide a simple interface
     to all sorts of things "that can be bumped".
   }
}
}

\section{Disclaimer}{

\itemize{
   \item{
   \strong{This package is really new. So please test it with
   repositories that are not crucial for your productive work!}
   }
}
}

\section{Inspiration and aknowledgements}{

This package was greatly inspired by these ressources:
\itemize{
   \item{\url{http://nvie.com/posts/a-successful-git-branching-model/}: } {
     A very thorough and well designed branching model.
   }
   \item{\url{https://gist.githubusercontent.com/pete-otaqui/4188238/raw/9675e75948294fd3b05186f30af6ff97e7e9e47e/bumpversion.sh}: } {
     Bash script for Git version bumping.
   }
   \item{\url{http://stackoverflow.com/questions/18648737/use-r-to-push-local-repo-to-github-on-windows}: } {
     Description of how to access GitHub repositories via R using HTTP
     authentication.
   }

}
}
\author{
Janko Thyson \email{janko.thyson@gmail.com}
}
\references{
\url{http://github.com/jankothyson/bumpr}
}
\seealso{
\url{http://git-scm.com}, \url{https://github.com/}
}

